demo/requirements.txt
chromadb==0.5.20 # Vector Database
sentence-transformers==3.3.1 # CrossEncoder Re-ranking
streamlit==1.40.1 # Application UI
PyMuPDF==1.24.14 # PDF Document loader
langchain-community==0.3.7 # Utils for text splitting
langchain-chroma==0.2.2
boto3>=1.37.1
streamlit>=1.40.1
streamlit-chat>=0.1.1
langchain-aws>=0.2.13
gensim>=4.3.3  # Play with text embedding models
pypdf>=5.3.0   # process pdf files
atlassian-python-api>=3.41.19  # Scrape atlassian confluence

